+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- DropDownSelect: Fix. Unable to choose 0 as a value from SuggestionList
 
 
 ## [0.27.0] - 2020-08-03