+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - DropDownSelect: Fix. Incorrect component behavior on Enter press. Should clear value, show placeholder and hide clear button.
 - DateTimePicker, TimePicker: Fix. Fixed timeMax timeMin limits. Time limiting works correctly.
 - DropDownSelect: Fix. Should select highlighted value on Enter press. Bug was actual when 'data' is an array of objects.
+- Input: Fix. Incorrect text input behavior with maxLength prop.
 
 
 ## [0.27.0] - 2020-08-03